카드 목록 페이지 - 카드 삭제 구현
