+++ younchong/src/App.jsx
@@ -13,7 +13,7 @@ function App() {
       <Routes>
         <Route path="/" element={<CardMaker setNewCard={setNewCard}/>} />
         <Route path="/list" element={<CardList cards={cards} setNewCard={setNewCard} setCard={setCard}/>} />
-        <Route path="/add" element={<CardAddPage card={newCard} setCard={setCard}/> } />
+        <Route path="/add" element={<CardAddPage card={newCard} setCard={setCard} setNewCard={setNewCard}/> } />
       </Routes>
     </BrowserRouter>
   );
