+++ younchong/src/components/card_add_page/add_button/add_button.module.css
@@ -1,0 +1,11 @@
+.button {
+  font-size: 1.5rem;
+  border: none;
+  background: white;
+}
+
+.button:hover {
+  transform: scale(1.1);
+  transition: 1s;
+  cursor: pointer;
+}
