+++ younchong/src/components/card_add_page/delete_button/delete_button.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import styles from "./delete_button.module.css";
+
+const DeleteButton = ({setNewCard, navigate}) => {
+  const onClick = (e) => {
+    e.preventDefault();
+    setNewCard(null);
+    navigate("/list");
+  }
+  return (
+    <button className={styles.button} onClick={onClick}>
+      X
+    </button>
+  );
+};
+
+export default DeleteButton;
