+++ younchong/src/components/card_add_page/delete_button/delete_button.module.css
@@ -1,0 +1,13 @@
+.button {
+  color: red;
+  font-size: 2rem;
+
+  border: none;
+  background: white;
+}
+
+.button:hover {
+  transform: scale(1.1);
+  transition: 1s;
+  cursor: pointer;
+}
