+++ younchong/src/components/card_make_page/card_cvc_input/card_cvc_input.module.css
@@ -45,9 +45,7 @@
 }
 
 .tooltip {
-  position: relative;
   display: block;
-
   
   color: #9ca3af;
   margin-left: 1rem;
